Fixing shadow not showing up
This commit aligned the code more with what was presented in the course and fixed the shadows not showing up.

+++ Projects/OpenGLCourseApp/OpenGLCourseApp/DirectionalLight.h
@@ -7,7 +7,7 @@ class DirectionalLight : public Light
     public:
         DirectionalLight();
 
-        DirectionalLight(GLfloat shadowWidth, GLfloat shadowHeight,
+        DirectionalLight(GLuint shadowWidth, GLuint shadowHeight,
             GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat dIntensity,
             GLfloat xDirection, GLfloat yDirection, GLfloat zDirection);
